add tests for the output writer
