+++ lib/output_writer.rb
@@ -1,11 +1,13 @@
 module Unipept
   class OutputWriter
+    attr_reader :output
+
     def initialize(file)
       @output = file ? File.open(file, 'a') : $stdout
     end
 
-    def write_line(string)
-      @output.write string
+    def write_line(line)
+      @output.write line
     end
   end
 end
